Renaming the test project and refactoring all test classes namespaces
