+++ DatahubTest/EmailNotificationTest.cs
@@ -13,7 +13,7 @@ using NRCan.Datahub.Shared.Services;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace DatahubTest
+namespace Datahub.Tests
 {
     public class EmailNotificationTestFixture: IDisposable
     {
