+++ DatahubTest/Metadata/MetadataDefinitionsTests.cs
@@ -4,7 +4,10 @@ using System;
 using System.Linq;
 using Xunit;
 
-namespace DatahubTest
+/// <summary>
+/// Cannot name it 'Datahub.Tests.Metadata' or will conflict with an existing class
+/// </summary>
+namespace Datahub.Tests.Meta_Data 
 {
     public class MetadataDefinitionsTests
     {
