+++ NRCan.Datahub.Portal.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.0.31710.8
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NRCan.Datahub.Portal", "WebPortal\NRCan.Datahub.Portal.csproj", "{7C55A80E-F420-45A7-A85E-BBF054533EE3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DatahubTest", "DatahubTest\DatahubTest.csproj", "{9437DD2A-0322-47ED-8FA2-96B8B87928CB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Datahub.Tests", "DatahubTest\Datahub.Tests.csproj", "{9437DD2A-0322-47ED-8FA2-96B8B87928CB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Datahub.Core", "Datahub.Core\Datahub.Core.csproj", "{E536F8EC-31D3-4C07-AEB2-C2A8D8634480}"
 EndProject
@@ -80,8 +80,8 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		RESX_ConfirmAddLanguageFile = True
-		RESX_AutoCreateNewLanguageFiles = True
 		SolutionGuid = {9DC4B901-16FF-4FF2-9ED6-B286EA17048E}
+		RESX_AutoCreateNewLanguageFiles = True
+		RESX_ConfirmAddLanguageFile = True
 	EndGlobalSection
 EndGlobal
